Fix warnings in the HIP backend

+++ core/unit_test/hip/TestHIP_UnifiedMemory_ZeroMemset.cpp
@@ -28,8 +28,8 @@ TEST(hip, unified_memory_zero_memset) {
       << "this test should only be run with HIP unified memory enabled";
 #endif
 
-  constexpr size_t N = 1024 * 1024;  // size doesn't matter
-  std::vector<int> v(N, 1);          // initialize to non-zero
+  constexpr size_t N = static_cast<size_t>(1024 * 1024);  // size doesn't matter
+  std::vector<int> v(N, 1);  // initialize to non-zero
   Kokkos::View<int*, Kokkos::HIPSpace> a(v.data(), N);
 
   // zero with deep_copy (this is where the error occurs)
